Move findKeywordInFirstParagraph to researches folder
Adjust the reference and delete it from language specific researches file
